TERMINADO, POR FIN

+++ TP1/TP 1/funciones.h
@@ -5,6 +5,6 @@
 int Sumar(int x, int y);
 int Restar(int x, int y);
 int Multiplicar(int x, int y);
-float Dividir(float x, float y);
+float Dividir(int x, int y);
 int Factorial(int numero);
 int PedirNumero(void);
